pyproject.toml
[tool.poetry]
name = "nomad-hsqdoc"
version = "0.1.0"
description = ""
authors = ["seedjin298 <seedmemoria@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-environ = "^0.11.2"
django = "^5.1.2"
djangorestframework = "^3.15.2"
pyjwt = "^2.9.0"
streamlit = "^1.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
